requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
astronomer-cosmos>=1.0.2
astro-sdk-python[amazon,snowflake]>=1.1.0
apache-airflow-providers-http==4.13.1
apache-airflow-providers-amazon==9.6.1
apache-airflow-providers-snowflake==5.7.0
apache-airflow-providers-common-sql==1.26.0
boto3>=1.37.0, <2
requests>=2.27.0, <3